refactor(price-adjustment-modal): enhance modal functionality and improve error handling refactor(adjustment-fields): change input step for adjust value to whole numbers refactor(date-selection): add cursor pointer to date range and specific dates labels refactor(filters-bar): simplify SelectTrigger and SelectItem components refactor(select): update cursor styles for select components refactor(usePriceAdjustments): streamline API calls and improve error messaging refactor(useRoomAvailability): optimize blocked dates management and authentication handling
