+++ apps/web/src/hooks/useRoomAvailability.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useCallback } from "react";
+import { useState, useCallback, useEffect } from "react";
 import { useSession } from "next-auth/react";
 import axios from "axios";
 import { toast } from "sonner";
@@ -10,177 +10,133 @@ import type {
   BlockDatesRequest,
   UnblockDatesRequest,
 } from "@/types/room";
-
-const createApiInstance = (accessToken?: string) => {
-  const api = axios.create({
-    baseURL: process.env.NEXT_PUBLIC_API_URL || "http://localhost:8000/api/v1",
-  });
-
-  if (accessToken) {
-    api.defaults.headers.common.Authorization = `Bearer ${accessToken}`;
-  }
-
-  return api;
-};
+import { api } from "@/lib/axios";
+import { setAuthToken } from "@/lib/axios";
 
 export function useRoomAvailability(roomId: string) {
-  const { data: session } = useSession();
+  const { data: session, status } = useSession();
   const [blockedDates, setBlockedDates] = useState<RoomAvailability[]>([]);
-  const [loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
+
+  const authToken = session?.user?.accessToken ?? null;
 
   const fetchBlockedDates = useCallback(
     async (startDate?: string, endDate?: string) => {
-      if (!session?.user?.accessToken || !roomId) return;
+      if (!authToken || !roomId) return;
 
       setLoading(true);
       setError(null);
 
       try {
-        const api = createApiInstance(session.user.accessToken);
         const params = new URLSearchParams();
         if (startDate) params.append("startDate", startDate);
         if (endDate) params.append("endDate", endDate);
 
-        const response = await api.get<RoomAvailabilityApiResponse>(
+        const res = await api.get<RoomAvailabilityApiResponse>(
           `/rooms/${roomId}/availability?${params.toString()}`
         );
-        console.log("Fetched blocked dates:", response.data.data);
-        setBlockedDates(response.data.data);
+        setBlockedDates(res.data.data);
       } catch (err) {
-        console.error("Error fetching blocked dates:", err);
         const errorMessage =
           axios.isAxiosError(err) && err.response?.data?.message
-            ? err.response.data.message
+            ? (err.response.data.message as string)
             : "Failed to fetch blocked dates";
+        console.error("fetchBlockedDates:", err);
         setError(errorMessage);
         toast.error(errorMessage);
       } finally {
         setLoading(false);
       }
     },
-    [roomId, session?.user?.accessToken]
+    [authToken, roomId]
   );
 
   const blockDates = useCallback(
     async (dates: string[]) => {
-      if (!session?.user?.accessToken || !roomId) {
-        toast.error("Authentication required");
-        return;
-      }
-
-      if (dates.length === 0) {
-        toast.error("No dates provided to block");
-        return;
-      }
+      if (!authToken) return toast.error("Authentication required");
+      if (dates.length === 0) return toast.error("No dates provided to block");
 
       try {
-        const api = createApiInstance(session.user.accessToken);
-        const response = await api.post<RoomAvailabilityApiResponse>(
+        const res = await api.post<RoomAvailabilityApiResponse>(
           `/rooms/${roomId}/block`,
           { dates } as BlockDatesRequest
         );
 
-        // Update local state with newly blocked dates
-        const newBlockedDates = response.data.data;
-        console.log("New blocked dates from API:", newBlockedDates);
+        const newBlocked = res.data.data;
         setBlockedDates((prev) => {
-          const updated = [...prev];
-
-          newBlockedDates.forEach((blocked) => {
-            const existingIndex = updated.findIndex(
-              (item) => item.date === blocked.date
-            );
-
-            if (existingIndex >= 0) {
-              updated[existingIndex] = blocked;
-            } else {
-              updated.push(blocked);
-            }
+          const merged = [...prev];
+          newBlocked.forEach((b) => {
+            const idx = merged.findIndex((m) => m.date === b.date);
+            if (idx >= 0) merged[idx] = b;
+            else merged.push(b);
           });
-
-          const result = updated.sort(
+          return merged.sort(
             (a, b) => new Date(a.date).getTime() - new Date(b.date).getTime()
           );
-          console.log("Updated blocked dates state:", result);
-          return result;
         });
 
         toast.success(
           `Blocked ${dates.length} date${dates.length > 1 ? "s" : ""}`
         );
-        return newBlockedDates;
+        return newBlocked;
       } catch (err) {
-        console.error("Error blocking dates:", err);
         const errorMessage =
           axios.isAxiosError(err) && err.response?.data?.message
-            ? err.response.data.message
+            ? (err.response.data.message as string)
             : "Failed to block dates";
+        console.error("blockDates:", err);
         toast.error(errorMessage);
         throw err;
       }
     },
-    [roomId, session?.user?.accessToken]
+    [authToken, roomId]
   );
 
   const unblockDates = useCallback(
     async (dates: string[]) => {
-      if (!session?.user?.accessToken || !roomId) {
-        toast.error("Authentication required");
-        return;
-      }
-
-      if (dates.length === 0) {
-        toast.error("No dates provided to unblock");
-        return;
-      }
+      if (!authToken) return toast.error("Authentication required");
+      if (dates.length === 0)
+        return toast.error("No dates provided to unblock");
 
       try {
-        const api = createApiInstance(session.user.accessToken);
         await api.post(`/rooms/${roomId}/unblock`, {
           dates,
         } as UnblockDatesRequest);
-
-        // Remove unblocked dates from local state
-        setBlockedDates((prev) =>
-          prev.filter((item) => !dates.includes(item.date))
-        );
-
+        setBlockedDates((prev) => prev.filter((p) => !dates.includes(p.date)));
         toast.success(
           `Unblocked ${dates.length} date${dates.length > 1 ? "s" : ""}`
         );
       } catch (err) {
-        console.error("Error unblocking dates:", err);
         const errorMessage =
           axios.isAxiosError(err) && err.response?.data?.message
-            ? err.response.data.message
+            ? (err.response.data.message as string)
             : "Failed to unblock dates";
+        console.error("unblockDates:", err);
         toast.error(errorMessage);
         throw err;
       }
     },
-    [roomId, session?.user?.accessToken]
+    [authToken, roomId]
   );
 
   const isDateBlocked = useCallback(
-    (date: string) => {
-      const result = blockedDates.some((item) => item.date === date);
-      console.log(`Checking if ${date} is blocked:`, result);
-      console.log(
-        "Available blocked dates:",
-        blockedDates.map((d) => d.date)
-      );
-      return result;
-    },
+    (date: string) => blockedDates.some((b) => b.date === date),
     [blockedDates]
   );
-
   const isDateAvailable = useCallback(
-    (date: string) => {
-      return !isDateBlocked(date);
-    },
+    (date: string) => !isDateBlocked(date),
     [isDateBlocked]
   );
+
+  useEffect(() => {
+    setAuthToken(authToken);
+    if (status === "authenticated" && authToken && roomId) {
+      fetchBlockedDates();
+    } else if (status !== "loading") {
+      setLoading(false);
+    }
+  }, [status, authToken, roomId, fetchBlockedDates]);
 
   return {
     blockedDates,
